dom/payments/test/test_shippingOptions.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1440041
https://bugzilla.mozilla.org/show_bug.cgi?id=1443914
-->
<head>
  <meta charset="utf-8">
  <title>Test for shippingOptions related bugs</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="./DefaultData.js"></script>
  <script type="application/javascript">

  "use strict";
  SimpleTest.waitForExplicitFinish();

  var gUrl = SimpleTest.getTestFileURL('ShippingOptionsChromeScript.js');
  var gScript = SpecialPowers.loadChromeScript(gUrl);

  function testFailHandler(message) {
    ok(false, message);
  }
  function testPassHandler(message) {
    ok(true, message);
  }
  gScript.addMessageListener("test-fail", testFailHandler);
  gScript.addMessageListener("test-pass", testPassHandler);

  async function requestChromeAction(action, params) {
    await new Promise(resolve => {
      gScript.addMessageListener(`${action}-complete`, function completeListener() {
        gScript.removeMessageListener(`${action}-complete`, completeListener);
        resolve();
      });
      gScript.sendAsyncMessage(action, params);
    });
  }

  let shippingOptions = [{
    id: "NormalShipping",
    label: "NormalShipping",
    amount: {
      currency: "USD",
      value: "10.00",
    },
    selected: true,
  },{
    id: "FastShipping",
    label: "FastShipping",
    amount: {
      currency: "USD",
      value: "5.00",
    },
    selected: false,
  }]

  // testing function main body
  async function testShippingOptionsTemplate(initDetails,
                                             optionUpdateDetails,
                                             testName,
                                             expectedRequestOption,
                                             expectedOptionChangeOption,
                                             expectedResponseOption) {
    info(testName);
    const expectedResults = {testName: testName,
                             requestResult: expectedRequestOption,
                             changeOptionResult: expectedOptionChangeOption,
                             responseResult: expectedResponseOption,};
    await requestChromeAction("set-expected-results", expectedResults);
    info("expected results setup finish");
    return new Promise(async (resolve) => {
      const request = new PaymentRequest(defaultMethods, initDetails, defaultOptions);
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      is(request.shippingOption, expectedRequestOption,
         `${testName}: request.shippingOption should be ${expectedRequestOption} after created.`);
      if (optionUpdateDetails) {
        request.addEventListener("shippingoptionchange", event => {
          is(request.shippingOption, expectedOptionChangeOption,
             `${testName}: request.shippingOption should be ${expectedOptionChangeOption} in shippingoptionchange event`);
          event.updateWith(optionUpdateDetails);
        });
      }
      try {
        const response = await request.show();
        is(response.shippingOption, expectedResponseOption,
           `${testName}: response.shippingOption should be ${expectedResponseOption}.`);
        await response.complete("success");
      } catch (error) {
        ok(false, `${testName}: unexpected error '${error.name}'.`);
      }
      await handler.destruct();
      resolve();
    });
  }

  // test no selected shipping option in default
  async function testNoSelectedShippingOptions() {
    return testShippingOptionsTemplate(defaultDetails,   // initial details
                                       null,             // update details for optionchange
                                       "testNoSelectedShippingOptions", // test name
                                       null,             // expected request.shippintOption after create
                                       null,             // expected request.shippingOption after optionchange
                                       null);            // expected response.shippingOption
  }

  // test select one shipping option in default
  async function testSelectedOneShippingOption() {
    let details = Object.assign({}, defaultDetails);
    details.shippingOptions = shippingOptions;
    details.shippingOptions[0].selected = true;
    details.shippingOptions[1].selected = false;
    const expectedOption = details.shippingOptions[0].id;
    return testShippingOptionsTemplate(details,          // initial details
                                       null,             // update details for optionchange
                                       "testSelectedOneShippingOption", // test name
                                       expectedOption,   // expected request.shippintOption after create
                                       null,             // expected request.shippingOption after optionchange
                                       expectedOption);  // expected response.shippingOption
  }

  // test select multiple shipping options in default
  async function testMultiSelectedShippingOptions() {
    let details = Object.assign({}, defaultDetails);
    details.shippingOptions = shippingOptions;
    details.shippingOptions[0].selected = true;
    details.shippingOptions[1].selected = true;
    const expectedOption = details.shippingOptions[1].id;
    return testShippingOptionsTemplate(details,          // initial details
                                       null,             // update details for optionchange
                                       "testMultiSelectedShippingOptions", // test name
                                       expectedOption,   // expected request.shippintOption after create
                                       null,             // expected request.shippingOption after optionchange
                                       expectedOption);  // expected response.shippingOption
  }

  // test no selected shipping option in default, but selected by user
  async function testSelectedByUser() {
    let updateDetails = Object.assign({}, defaultDetails);
    updateDetails.shippingOptions = shippingOptions;
    updateDetails.shippingOptions[0].selected = true;
    updateDetails.shippingOptions[1].selected = false;
    const expectedOption = updateDetails.shippingOptions[0].id;
    return testShippingOptionsTemplate(defaultDetails,   // initial details
                                       updateDetails,    // update details for optionchange
                                       "testSelectedByUser", // test name
                                       null,             // expected request.shippintOption after create
                                       expectedOption,   // expected request.shippingOption after optionchange
                                       expectedOption);  // expected response.shippingOption
  }

  // test no selected shipping option in default, but selected by user then updated
  // by merchant to the other.
  async function testUpdateSelectedByMerchant() {
    let updateDetails = Object.assign({}, defaultDetails);
    updateDetails.shippingOptions = shippingOptions;
    updateDetails.shippingOptions[0].selected = false;
    updateDetails.shippingOptions[1].selected = true;
    const expectedOption = updateDetails.shippingOptions[0].id;
    const expectedResponse = updateDetails.shippingOptions[1].id;
    return testShippingOptionsTemplate(defaultDetails,   // initial details
                                       updateDetails,    // update details for optionchange
                                       "testUpdateSelectedByMerchant", // test name
                                       null,             // expected request.shippintOption after create
                                       expectedOption,   // expected request.shippingOption after optionchange
                                       expectedResponse);// expected response.shippingOption
  }

  // test update shipping options to null
  async function testUpdateShippingOptionsToNull() {
    let updateDetails = Object.assign({}, defaultDetails);
    delete updateDetails.shippingOptions;
    const expectedOption = defaultDetails.shippingOptions[0].id;
    return testShippingOptionsTemplate(defaultDetails,   // initial details
                                       updateDetails,    // update details for optionchange
                                       "testUpdateShippingOptionsToNull", // test name
                                       null,             // expected request.shippintOption after create
                                       expectedOption,   // expected request.shippingOption after optionchange
                                       null);            // expected response.shippingOption
  }

  async function testUpdateWithSelectedShippingOption() {
    let initDetails = Object.assign({}, defaultDetails);
    initDetails.shippingOptions = shippingOptions;
    const expectedOption = initDetails.shippingOptions[1].id;
    const expectedResults = {testName: "testUpdateWithSelectedShippingOption",
                             requestResult: expectedOption,
                             changeOptionResult: expectedOption,
                             responseResult: expectedOption};
    await requestChromeAction("set-expected-results", expectedResults);
    return new Promise(async (resolve) => {
      const request = new PaymentRequest(defaultMethods, initDetails, defaultOptions);
      request.addEventListener("shippingoptionchange", async (event) => {
        ok(false, "Expected no event emitted, but got one.");
        try {
          request.abort();
        } catch (error) {
          ok(false, `Unexpected error '${error.name}' while aborting PaymentRequest.`);
        }
        resolve();
      });
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      try {
        const response = await request.show();
        await response.complete();
      } catch (error) {
        ok(false, `Unexpected error '${error.name}' while aborting PaymentRequest.`);
      }
      await handler.destruct();
      resolve();
    });
  }

  async function teardown() {
    gScript.addMessageListener("teardown-complete", function teardownCompleteHandler() {
      gScript.removeMessageListener("teardown-complete", teardownCompleteHandler);
      gScript.removeMessageListener("test-fail", testFailHandler);
      gScript.removeMessageListener("test-pass", testPassHandler);
      gScript.destroy();
      SimpleTest.finish();
    });
    gScript.sendAsyncMessage("teardown");
  }

  async function runTests() {
    try {
      await testNoSelectedShippingOptions()
      await testSelectedOneShippingOption();
      await testMultiSelectedShippingOptions();
      await testSelectedByUser();
      await testUpdateSelectedByMerchant();
      await testUpdateShippingOptionsToNull();
      await testUpdateWithSelectedShippingOption();
      await teardown();
    } catch(e) {
      ok(false, `test_shippingOpions.html: Unexpected error: ${e.name}`);
      SimpleTest.finish();
    }
  }

  window.addEventListener('load', function() {
    SpecialPowers.pushPrefEnv({
      'set': [
        ['dom.payments.request.enabled', true],
      ]
    }, runTests);
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1440041">Mozilla Bug 1440041</a>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1443914">Mozilla Bug 1443914</a>
</body>
</html>
